Drag Bibtex Files, restructuring
Major update.  Usage mode changed from loading json file (still
supported for initial graph) to drag-and-dropping a BibTex file directly
into the browser.

Restructured more code to get rid of the json loading and change_mode
stuff in the future.

Replaced the glowing background rectangle with a small colored ring
indicating the state of the paper.